renames/ deletes governance, bounties tests
